requirements.txt
fastapi
uvicorn[standard]
httpx
aiofiles
python-dotenv
sqlalchemy
python-multipart
pydantic